KAFKA-17532: Moved ShareGroupConfig to share module
