+++ build.gradle
@@ -889,7 +889,9 @@ project(':server') {
     implementation project(':metadata')
     implementation project(':server-common')
     implementation project(':storage')
+    implementation project(':share')
     implementation project(':group-coordinator')
+    implementation project(':coordinator-common')
     implementation project(':transaction-coordinator')
     implementation project(':raft')
     implementation libs.metrics
@@ -954,6 +956,8 @@ project(':share') {
 
   dependencies {
     implementation project(':server-common')
+    implementation project(':clients')
+    implementation project(':coordinator-common')
 
     implementation libs.slf4jApi
