+++ checkstyle/import-control.xml
@@ -248,12 +248,17 @@
 
   <subpackage name="server">
     <allow pkg="org.apache.kafka.common" />
+    <allow pkg="org.apache.kafka.coordinator.common" />
 
     <!-- This is required to make AlterConfigPolicyTest work. -->
     <allow pkg="org.apache.kafka.server.policy" />
 
     <subpackage name="telemetry">
       <allow class="org.apache.kafka.server.authorizer.AuthorizableRequestContext" />
+    </subpackage>
+
+    <subpackage name="share">
+      <allow pkg="org.apache.kafka.coordinator.common" />
     </subpackage>
 
   </subpackage>
@@ -342,6 +347,10 @@
       <allow pkg="org.jfree"/>
       <allow pkg="javax.imageio" />
       <allow pkg="scala" />
+    </subpackage>
+
+    <subpackage name="test">
+      <allow pkg="org.apache.kafka.coordinator.common"/>
     </subpackage>
   </subpackage>
